Minor refactoring and comments.

+++ plugins/shaders/Doom3ShaderLayer.cpp
@@ -192,7 +192,7 @@ ShaderLayer::CubeMapMode Doom3ShaderLayer::getCubeMapMode() const
     return _cubeMapMode;
 }
 
-double Doom3ShaderLayer::getAlphaTest() const
+float Doom3ShaderLayer::getAlphaTest() const
 {
     return _registers[_alphaTest];
 }
